fix VC++ compile issue
real_reql_cluster_interface.cc(55): error C2664: 'ql::changefeed::client_t::client_t(const ql::changefeed::client_t &)': cannot convert argument 2 from 'real_reql_cluster_interface_t::<lambda_56eb3f4f88c05fc89428f33402c85b51>' to 'const std::function<namespace_interface_access_t (const namespace_id_t &,signal_t *)> &'
