+++ src/rdb_protocol/real_table.hpp
@@ -9,6 +9,7 @@
 
 #include "rdb_protocol/configured_limits.hpp"
 #include "rdb_protocol/context.hpp"
+#include "rdb_protocol/namespace_interface_access.hpp"
 #include "rdb_protocol/protocol.hpp"
 
 namespace ql {
@@ -21,34 +22,6 @@ class client_t;
 /* `real_table_t` is a concrete subclass of `base_table_t` that routes its queries across
 the network via the clustering logic to a B-tree. The administration logic is responsible
 for constructing and returning them from `reql_cluster_interface_t::table_find()`. */
-
-/* `namespace_interface_access_t` is like a smart pointer to a `namespace_interface_t`.
-This is the format in which `real_table_t` expects to receive its
-`namespace_interface_t *`. This allows the thing that is constructing the `real_table_t`
-to control the lifetime of the `namespace_interface_t`, but also allows the
-`real_table_t` to block it from being destroyed while in use. */
-class namespace_interface_access_t {
-public:
-    class ref_tracker_t {
-    public:
-        virtual void add_ref() = 0;
-        virtual void release() = 0;
-    protected:
-        virtual ~ref_tracker_t() { }
-    };
-    namespace_interface_access_t();
-    namespace_interface_access_t(namespace_interface_t *, ref_tracker_t *, threadnum_t);
-    namespace_interface_access_t(const namespace_interface_access_t &access);
-    namespace_interface_access_t &operator=(const namespace_interface_access_t &access);
-    ~namespace_interface_access_t();
-
-    namespace_interface_t *get();
-
-private:
-    namespace_interface_t *nif;
-    ref_tracker_t *ref_tracker;
-    threadnum_t thread;
-};
 
 namespace ql {
 class reader_t;
